Add a new caption role
Add the new role as a subclass of sectionhead. At the moment, caption only
has table as its required context role. There was discussion on the mailing
list about using the caption role for figure captions as well. This seems
like a good idea (to me) but probably requires further discussion. I'm also
operating under the assumption that grids and treegrids do not need captions.
We can discuss this as a group.